+++ checks/check.command.go
@@ -23,7 +23,8 @@ but arguments are kept intact (quoted) when passed to the program.
     If the program is on your PATH, it will be found by its name alone.
   - args: An array of strings containing any arguments to pass.
     Optional.
-  - dir: TODO.
+  - dir: A string with the path in which to run the program.
+    Optional, defaults to the current directory.
 */
 type CheckCommand struct{}
 
@@ -33,6 +34,10 @@ func (c CheckCommand) Run(p Params) CheckResult {
 	if len(p.GetStrings("args")) > 0 {
 		args = p.GetStrings("args")
 	}
+	dir := p.GetString("dir")
+	if dir == "" {
+		// Already handled well by package cmd
+	}
 
 	line := program
 	if len(args) > 0 {
@@ -41,6 +46,7 @@ func (c CheckCommand) Run(p Params) CheckResult {
 	}
 
 	cmd := exec.Command(program, args...)
+	cmd.Dir = dir
 	err := cmd.Start()
 	if err != nil {
 		return newBasicResult(false, fmt.Sprintf("%s: %s", line, err.Error()))